+++ Innowise.Music/Services/ISearchService.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+using Innowise.Music.Model;
+
+namespace Innowise.Music.Services;
+
+public interface ISearchService
+{
+    Task<UnifiedSearchResponse?> UnifiedSearchAsync(string query);
+}